+++ src/main/java/io/taylor/wantedpreonboardingchallengebackend20/dto/response/ProductOrderResponse.java
@@ -1,4 +1,6 @@
 package io.taylor.wantedpreonboardingchallengebackend20.dto.response;
 
-public record ProductOrderResponse(long price, long quantity) {
+import java.math.BigDecimal;
+
+public record ProductOrderResponse(BigDecimal price, int quantity) {
 }
